Fixed some remarks
