+++ grails-app/controllers/com/secured/api/ApiTestController.groovy
@@ -6,6 +6,7 @@ import com.secured.auth.User
 import grails.converters.JSON
 import grails.plugin.springsecurity.annotation.Secured
 import grails.rest.RestfulController
+import org.grails.web.json.JSONElement
 
 @Secured(['permitAll'])
 class ApiTestController extends RestfulController<User> {
@@ -16,19 +17,25 @@ class ApiTestController extends RestfulController<User> {
 
     def index()
     {
-        def admin = User.findByUsername("admin")
-        def barcode = new BarCode(code: "test")
-        def company = new Company(address: "New York")
-        barcode.save()
-        company.addToBarCodes(barcode)
-        company.save()
-        admin.addToCompanies(company)
+        def admin = queryForResource(1)
+        def company = Company.findOrSaveWhere(address: "New York", user: admin)
+        def barcode = BarCode.findOrSaveWhere(code: "test", company: company)
         admin.save(true)
+
+        JSON jsonobject = new JSON()
+        jsonobject = admin as JSON
 
         withFormat {
             json{
-                render admin.companies[0].barCodes as JSON
+                render jsonobject
             }
         }
+    }
+
+    @Override
+    protected User queryForResource(Serializable id) {
+        User.where {
+            id == id
+        }.find()
     }
 }
