+++ grails-app/services/com/secured/auth/UserValidatorService.groovy
@@ -1,12 +1,13 @@
 package com.secured.auth
 
+import com.secured.auth.commands.UserCommand
 import grails.gorm.transactions.Transactional
 
 
 @Transactional
 class UserValidatorService extends PatternValidatorService {
 
-    def isPasswordValid(User usr)
+    def isPasswordValid(UserCommand usr)
     {
 
         if(usr == null)
@@ -16,21 +17,25 @@ class UserValidatorService extends PatternValidatorService {
         {
             case -2:
                 usr.errors.rejectValue("password","user.password.tooshort")
+                usr.password = ""
                 return false
             case -1:
                 usr.errors.rejectValue("password","user.password.toolong")
+                usr.password = ""
                 return false
             case 0:
                 usr.errors.rejectValue("password","user.password.whitespaces")
+                usr.password = ""
                 return false
             case 1:
                 usr.errors.rejectValue("password","user.password.tooweak")
+                usr.password = ""
                 return false
         }
         return true
     }
 
-    def isUsernameValid(User usr)
+    def isUsernameValid(UserCommand usr)
     {
         if(usr == null)
             return false
