+++ src/test/groovy/com/secured/auth/commands/UserCommandSpec.groovy
@@ -1,0 +1,18 @@
+package com.secured.auth.commands
+
+import grails.testing.gorm.DomainUnitTest
+import spock.lang.Specification
+
+class UserCommandSpec extends Specification implements DomainUnitTest<UserCommand> {
+
+    def setup() {
+    }
+
+    def cleanup() {
+    }
+
+    void "test something"() {
+        expect:"fix me"
+            true == false
+    }
+}
